Update CNAME
This commit adds a CNAME file to the root directory of the repository. The purpose of this file is to link the custom domain www.cimarron.live to the GitHub Pages site hosted within the repository cimarron-cuban/gpodder.github.io. By adding this file, we ensure that the custom domain correctly points to the GitHub Pages site, providing a seamless experience for visitors accessing the website.